+++ core/mount/mount_darwin.go
@@ -1,5 +1,3 @@
-//go:build !windows && !linux
-
 /*
    Copyright The containerd Authors.
 
@@ -16,19 +14,11 @@
    limitations under the License.
 */
 
-package os
+package mount
 
-import (
-	"github.com/containerd/containerd/v2/core/mount"
-	"github.com/containerd/errdefs"
-)
+import "github.com/containerd/errdefs"
 
-// Mount will call unix.Mount to mount the file.
-func (RealOS) Mount(source string, target string, fstype string, flags uintptr, data string) error {
+// Mount to the provided target.
+func (m *Mount) mount(target string) error {
 	return errdefs.ErrNotImplemented
-}
-
-// Unmount will call Unmount to unmount the file.
-func (RealOS) Unmount(target string) error {
-	return mount.Unmount(target, 0)
 }
